Remove message for existing 'users' table
The code that checks if the 'users' table already exists and prints a message if it does has been removed. Specifically, the lines `} else {` and `echo "La table 'users' existe déjà.<br>";` have been deleted. This means that the script will no longer output a message when the 'users' table already exists.